% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BadTrialsEyeTrack.R
\name{BadTrialsEyeTrack}
\alias{BadTrialsEyeTrack}
\title{Remove BadTrials from EyeTracking Data}
\usage{
BadTrialsEyeTrack(SubDat, stimData, nSub, nTrials, order)
}
\description{
\code{BadTrialsEyeTrack} uses information from Raw Data and finds trials that satisfy certain criteria
In current implementation, it finds raw values that are at 0,0 or negative x, negative y values
It uses a rough threshold of 50 such values (can be changed)
in addition, it needs stim file to find tracking ratio and uses a rough criteria of 97% tracking ratio.
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

