% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyFun.R
\name{classifyFun}
\alias{classifyFun}
\title{simple function to create permutation testing of a classifier}
\usage{
classifyFun(classifierName, genclassifier, Data, predictorCol, selectedCols,
  ranges = NULL, tune = FALSE, cost = 1, gamma = 0.5, ...)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

