% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dist_3d}
\alias{dist_3d}
\title{3d distance}
\usage{
dist_3d(p1, p2)
}
\arguments{
\item{p1}{numeric vector: length of 3}

\item{p2}{numeric vector: length of 3}
}
\value{
numeric: 3d distance between p1 and p2
}
\description{
3d distance
}
