% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dot}
\alias{dot}
\title{Dot product}
\usage{
dot(u, v)
}
\arguments{
\item{u}{Numeric vector}

\item{v}{Numeric vector}
}
\value{
Numeric dot product of `u` and `v`.
}
\description{
Dot product
}
