% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duty_factor_calc.R
\name{duty_factor_calc}
\alias{duty_factor_calc}
\title{Calculate duty factor}
\usage{
duty_factor_calc(M, thresh, plot_strides = FALSE)
}
\arguments{
\item{M}{data.frame:}

\item{thresh}{numeric: Threshold for considering the foot off or on the substrate}

\item{plot_strides}{boolean: Plot the strides}
}
\value{
list
\enumerate{
  \item \code{n_strides_df} data.frame: number of strides
  \item \code{duty_factor} numeric: mean duty factor
}
}
\description{
Calculate duty factor
}
