% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EyeResults.R
\name{EyeResults}
\alias{EyeResults}
\title{Function to save results for AOI fixation}
\usage{
EyeResults(EyeDataHand, EyeDataBottle, order)
}
\description{
Specific for AOIs of Hand and Bottle for Reading intention experiment
Assumes EyeData has columns of , intention,Subject.No or Subject.Name,
Data columns are Net.Dwell.Time..ms., Entry.Time..ms.
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

