% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinematicAnalysis-package.R
\docType{package}
\name{KinematicAnalysis-package}
\alias{KinematicAnalysis-package}
\alias{KinematicAnalysis}
\title{KinematicAnalysis}
\description{
More description
}
\details{
\tabular{ll}{ Package: \tab KinematicAnalysis\cr
              Type: \tab Package\cr
              Version: \tab 0.2.0\cr
              Date: \tab 2022-03-15\cr
              License: \tab MIT\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{package}
