% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadAOIData.R
\name{LoadAOIData}
\alias{LoadAOIData}
\title{Function to load AOI Data from BeGaze (exported as csv file)}
\usage{
LoadAOIData()
}
\description{
Function to load AOI Data from BeGaze (exported as csv file)
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

