% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LoadRawData.R
\name{LoadRawData}
\alias{LoadRawData}
\title{Function to load the Raw Data exported from BeGaze software.}
\usage{
LoadRawData(delim)
}
\description{
This is necessary for removal of the outlier trials.
This feeds data to the wrongTrials.R function
This also assumes that you export data as comma separated file in BeGaze.
Only input is the delimiter in the file after which the subject name or subject No.
(or both) are present. This in brief is the last portion of the name of the BeGaze experiment.
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

