% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadStimData.R
\name{LoadStimData}
\alias{LoadStimData}
\title{Function to load stim data containing Tracking ratio from BeGaze}
\usage{
LoadStimData()
}
\description{
Function to load stim data containing Tracking ratio from BeGaze
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

