% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_xyz_matrix}
\alias{make_xyz_matrix}
\title{Convert wide data to (x, y, z) rows}
\usage{
make_xyz_matrix(d)
}
\arguments{
\item{d}{Data.frame with points in wide format}
}
\value{
Data.frame with (x, y, z) points as rows 1, 2, and 3
}
\description{
Convert wide data to (x, y, z) rows
}
