% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/meanLocationIntervals.R
\name{meanLocationIntervals}
\alias{meanLocationIntervals}
\title{Function to get mean values of X and Y co-ordinates after time = time for each subject}
\usage{
meanLocationIntervals(RawData, timeInter, order)
}
\description{
Assumes the rawData has columns = X,Y and NormalisedTime
it also has a column Subject
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

