% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanLocationT.R
\name{meanLocationT}
\alias{meanLocationT}
\title{Function to get mean values of X and Y co-ordinates after time = time for each subject}
\usage{
meanLocationT(RawData, time, order)
}
\description{
Assumes the rawData has columns = X,Y and NormalisedTime
it also has a column Subject
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

