% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeExcelData.R
\name{mergeExcelData}
\alias{mergeExcelData}
\title{Read Data from multiple files/sheets of Excel files and create a single dataframe}
\usage{
mergeExcelData(filenames, SheetNo, selectedData = NULL, ...)
}
\description{
\code{mergeExcelData} takes in multiple filenames/single excel file with multiple sheets and merges as
a single dataframe.
uses packages: openxlsx
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
The function assumes that the files to be merged have the same number of columns
(necessity for rbind)
}

