% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalisedTime.R
\name{normalisedTime}
\alias{normalisedTime}
\title{Function to get normalised parameters from Eye tracking time series raw data}
\usage{
normalisedTime(SubData, stimData, order)
}
\description{
Needs the following:
:
\enumerate{
\item SubDat
  \enumerate{
     \item Subject.(Name or No)
     \item Trial
     \item Time (has MSG as one value)
   \item X and Y gaze positions
}
 \item stimData
   \enumerate{
   \item Duration
}
}
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

