% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{plot3d}
\alias{plot3d}
\title{Plot point in 3D with plotly}
\usage{
plot3d(M, cal_rotate)
}
\arguments{
\item{M}{Data.frame of points in wide format}

\item{cal_rotate}{Calibration points}
}
\description{
Plot point in 3D with plotly
}
