% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferred_speed.R
\name{preferred_speed}
\alias{preferred_speed}
\title{Calculate preferred speed}
\usage{
preferred_speed(M)
}
\arguments{
\item{M}{data.frame: Coordinates for markers}
}
\value{
numeric: Mean forward velocity
}
\description{
Calculate preferred speed
}
