% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_kine_data.R
\name{process_kine_data}
\alias{process_kine_data}
\title{Process kinematic data}
\usage{
process_kine_data(Trial_Info, ii, cal_coords)
}
\arguments{
\item{Trial_Info}{data.frame with information about trials}

\item{ii}{integer index for the row of Trial_Info to analyze}

\item{cal_coords}{Corrdinates for calibration object}
}
\value{

}
\description{
Process kinematic data
}
