% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_video.R
\name{process_video}
\alias{process_video}
\title{Process video file}
\usage{
process_video(ii, Trial_Info, bef_aft, calibration_coords)
}
\arguments{
\item{ii}{integer: row od \code{Trial_Info} to process}

\item{Trial_Info}{data.frame: Structure containing all the trials to process}
}
\value{
data.frame
\enumerate{
    \item \code{MouseID}
    \item \code{Trial}
    \item \code{Before_After}
    \item \code{mean_speed}
    \item \code{stride_frequency}
    \item \code{mean_stride_length}
    \item \code{n_strides_df}
    \item \code{duty_factor}
    \item \code{stance_width}
    \item \code{sacrum_vert_displacement}
}
}
\description{
Process video file
}
