% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_kine_functions.R
\name{read_xyz}
\alias{read_xyz}
\title{Read in 3d points from DeepLabCut}
\usage{
read_xyz(f, frame_start, frame_end, conv = 0.1004758)
}
\arguments{
\item{conv}{}
}
\value{

}
\description{
Read in 3d points from DeepLabCut
}
