% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBehavData.R
\name{readBehavData}
\alias{readBehavData}
\title{Read Data from multiple sheets of an Excel file}
\usage{
readBehavData(SheetName, SheetNo)
}
\description{
Read Data from multiple sheets of an Excel file
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

