% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodrigues_rotation.R
\name{rodrigues_rotation}
\alias{rodrigues_rotation}
\title{Calculate Rodrigues rotation matrix}
\usage{
rodrigues_rotation(v_zero, v_axis, verbose = FALSE)
}
\arguments{
\item{v_zero}{Numeric vector of axis to zero}

\item{v_axis}{Numeric vector of axis to rotate about}

\item{verbose}{Logical indicating whether to print message about rotation
angle}
}
\value{
Matrix of Rodrigues rotation parameters
}
\description{
Calculate Rodrigues rotation matrix
}
