% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_landmark_points.R
\name{rotate_landmark_points}
\alias{rotate_landmark_points}
\title{Rotate landmark points}
\usage{
rotate_landmark_points(M, rot_mat)
}
\arguments{
\item{M}{data.frame of landmark points}

\item{rot_mat}{3x3 numeric rotation matrix}
}
\value{
M points rotated using rot_mat
}
\description{
Given a data.frame of landmarks in x, y, z, x, y, z... configuration
and a rotation matrix, rotate the points using the matrix and
reconstruct the original format.
}
\details{
Currently this function assumes that there will be points for "eye",
"sacrum", "hip", "knee", "ankle", and "foot".
}
