% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_points.R
\name{rotate_points}
\alias{rotate_points}
\title{Rotate a wide set of points}
\usage{
rotate_points(x, Rx, Ry)
}
\arguments{
\item{x}{Data.frame or matrix of points in wide form}

\item{Rx}{Rotation matrix for the x-axis}

\item{Ry}{Rotation matrix for the y-axis}
}
\value{
Data.frame with rotated points
}
\description{
Rotate a wide set of points
}
