% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_kine_functions.R
\name{rotation_matrices}
\alias{rotation_matrices}
\title{Calculate rotation matrices for a set of calibration coordinates}
\usage{
rotation_matrices(cal_coords)
}
\arguments{
\item{cal_coords}{}
}
\value{

}
\description{
Calculate rotation matrices for a set of calibration coordinates
}
