% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selectTrial.R
\name{selectTrial}
\alias{selectTrial}
\title{Function to select a  specific trial from raw or data containing subject and Trial information}
\usage{
selectTrial(Dataframe, Subject, trial)
}
\description{
Function to select a  specific trial from raw or data containing subject and Trial information
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

