% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stride_parameters.R
\name{stride_parameters}
\alias{stride_parameters}
\title{Calculate stride parameters}
\usage{
stride_parameters(ii, M)
}
\arguments{
\item{ii}{integer: Row index of Trials_Info to process}

\item{M}{data.frame: Coordinates for markers}
}
\value{
list
\enumerate{
  \item \code{stride_frequency} numeric: mean stride frequency
  \item \code{mean_stride_length} numeric: mean stride length
}
}
\description{
Calculate stride parameters
}
