% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_points.R
\name{translate_points}
\alias{translate_points}
\title{Translate points}
\usage{
translate_points(d, origin)
}
\arguments{
\item{d}{Data.frame with points as columns (e.g., V1_x, V1_y, V1_z, V2_x,
V2_y, V2_z...)}

\item{origin}{Numeric vector with coordinates to translate points `d` to.}
}
\value{
Data.frame with points as columns translated.
}
\description{
Translate points
}
