% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_angle.R
\name{vector_angle}
\alias{vector_angle}
\title{Vector angle}
\usage{
vector_angle(A, B)
}
\arguments{
\item{A}{numeric vector}

\item{B}{numeric vector}
}
\value{
angle between A and B
}
\description{
Use the Law of Cosines to calculate the 3d angle between two vectors.
}
\examples{
vector_angle(c(0, 0, 1), c(0, 1, 0))

}
