% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyz_rotation_matrix.R
\name{xyz_rotation_matrix}
\alias{xyz_rotation_matrix}
\title{Calculate rotation matrix}
\usage{
xyz_rotation_matrix(theta, axis)
}
\arguments{
\item{theta}{numeric angle in radians}

\item{axis}{character for axis of rotation, "x", "y", or "z"}
}
\value{
3x3 rotation matrix
}
\description{
For an given theta and axis for rotation, return the appropriate 3x3
rotation matrix.
}
