% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_behav.R
\name{analyse_behav}
\alias{analyse_behav}
\title{Analyse behavioral Data from Eprime/text/Excel files specifying single or multiple accuracy columns}
\usage{
analyse_behav(resultfiles, scriptfile, signal, yesResp, save, integration,
  ResultCat = NULL)
}
\description{
\code{analyse_behav} uses takes information from a named text file and extracts multiple behavioral measures
In current implementation, it finds raw values that are at 0,0 or negative x, negative y values
uses packages: openxlsx, plyr
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

