% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinematicAnalysis-package.R,
%   R/KinematicAnalysis.R
\docType{package}
\name{KinematicAnalysis-package}
\alias{KinematicAnalysis}
\alias{KinematicAnalysis-package}
\title{KinematicAnalysis: 3D kinematic data analysis}
\description{
The package is used to aid with 3D kinematic data analysis.

More description
}
\details{
\tabular{ll}{ Package: \tab KinematicAnalysis\cr
              Type: \tab Package\cr
              Version: \tab 0.1\cr
              Date: \tab 2020-02-15\cr
              License: \tab MIT\cr
              LazyLoad: \tab yes\cr
              LazyData: \tab yes\cr }
}
\author{
Kevin M. Middleton (\url{middletonk@missouri.edu})
}
\keyword{internal}
\keyword{package}
