% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_kine_data.R
\name{process_kine_data}
\alias{process_kine_data}
\title{Process kinematic data}
\usage{
process_kine_data(Trial_Info, ii, calibration_coords, bef_aft)
}
\arguments{
\item{Trial_Info}{data.frame with information about trials}

\item{ii}{integer index for the row of Trial_Info to analyze}

\item{calibration_coords}{Coordinates for calibration object}

\item{bef_aft}{String for "Before" or "After"}
}
\value{

}
\description{
Process kinematic data
}
