% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Radians to degrees}
\usage{
rad2deg(r)
}
\arguments{
\item{r}{numeric vector of radians}
}
\value{
vector with r converted to degrees
}
\description{
Radians to degrees
}
