% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xyz.R
\name{read_xyz}
\alias{read_xyz}
\title{Read in 3d points from DeepLabCut}
\usage{
read_xyz(f, frame_start, frame_end, conv = 0.1004758)
}
\arguments{
\item{f}{String of file name to read}

\item{frame_start}{Integer of frame to start reading from}

\item{frame_end}{Integer of frame to end reading from}

\item{conv}{Numeric value for conversion from pixels to mm}
}
\value{
Tibble with time and (x, y, z) coordinates
}
\description{
Read in 3d points from DeepLabCut
}
