% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_kine_functions.R
\name{rodrigues_rotation}
\alias{rodrigues_rotation}
\title{Calculate Rodrigues rotation matrix}
\usage{
rodrigues_rotation(v_zero, v_axis, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Calculate Rodrigues rotation matrix
}
