% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotation_matrices.R
\name{rotation_matrices}
\alias{rotation_matrices}
\title{Calculate rotation matrices for a set of calibration coordinates}
\usage{
rotation_matrices(cal_coords)
}
\arguments{
\item{cal_coords}{Coordinates of calibration object in wide format.}
}
\value{
List containing:
`Rx` and `Ry` Rodrigues rotation matrices to zero about the x and y axes
`cal_rot` Calibration points rotated about x and y
}
\description{
Calculate rotation matrices for a set of calibration coordinates
}
