% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sigDetect.R
\name{sigDetect}
\alias{sigDetect}
\title{Part of Data Analysis behavioral Data from Eprime/text/Excel files specifying single or multiple accuracy columns}
\usage{
sigDetect(Data, signal, yesResp, condition, respSlides)
}
\description{
\code{sigDetect} calculates signal detection measures from the input dataframe
its necessary to have columns of participant response (MovieStart.RESP and ITI.Resp), condition column (Condizione)
and Confidence
}
\details{
The current setup is based on assumption that the participants can respond in two slides - respSlides[1] and respSlides[2]
if this is not the case, it goes to NA

Inputs:
       Data - Data frame that is input with columns of response, condition and confidence
       signal - condition that will be used as the signal
       condition - name of the column that has the condition
       respSlides - response slides where participants can respond (in Eprime e,g,)

Output:
      Data - dataframe with columns of Dprime and roc confidence

Currently,
}
\author{
Atesh Koul, RBCS, Istituto Italiano di technologia

\email{atesh.koul@gmail.com}
}

